requirements.txt
# ==================== Web Framework ====================
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# ==================== Database ====================
sqlalchemy>=2.0.23
alembic>=1.12.1
mysql-connector-python>=8.2.0

# ==================== Search & Storage ====================
# Azure AI Search
azure-search-documents>=11.4.0
azure-core>=1.26.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Storage and caching
redis>=5.0.1

# ==================== Message Queue ====================
kafka-python>=2.0.2

# ==================== AI & Embedding ====================
google-generativeai==0.3.1

# ==================== NLP & Text Processing ====================
jieba>=0.42.1
spacy>=3.7.2
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
nltk==3.8.1
numpy>=1.24.0

# ==================== Document Parsing ====================
pypdf>=3.17.0
python-docx>=1.1.0

# ==================== Security ====================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ==================== Utilities ====================
websockets>=12.0

# ==================== Optional RAG Evaluation ====================
# Azure AI Evaluation SDK (LLM-as-judge)
azure-ai-evaluation>=1.0.0

# Optional alternative evaluation stack
# ragas>=0.1.12
# datasets>=2.19.0
